# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: License Trading Company API
  description: A web application for customers of the company to create applications, which are then reviewed by staff members and eventually approved or rejected.
  contact:
    name: Petar Drumev
    email: drumevp@coventry.ac.uk
  license: 
    name: Apache 2.0,
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://opera-ski-3000.codio-box.uk/api/'
    description: Development server
...
paths:
  /users:
    summary: Represents a user

    get:
      tags: 
      - Users
      description: Returns list of all users.
      responses: 
        '200': 
          description: Users obtained
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/listUser
        '401':
          description: Not logged in
        '403':
          description: Only admins have access to see a list of all users
  /users/{id}:
    summary: Represents an individual user.
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric ID of the user.
      
    get:
      tags: 
      - Users
      description: Returns a single user.
      responses:
        '200':
          description: User obtained
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/viewUser
        '401':
          description: Not logged in
        '403':
          description: Non-admins only have access to their own profile.
        '404':
          description: Not found

    put:
      tags: 
      - Users
      description: Update a user record.
      requestBody:
        description: Data the user wants to change.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/updateUser
      responses:
        '201':
          description: User updated.
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/updatedUser
        '401':
          description: Not logged in
        '403':
          description: Non-admin users can only update their own account
        '404':
          description: Not found


    del:
      tags: 
      - Users
      description: Delete a user record
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/deletedUser
        '401':
          description: Not logged in
        '403':
          description: Non-admins can only delete their own accounts.
        '404':
          description: Not found
...
  /users/login:
    summary: Authenticate user credentials for access to API operations

    post:
      tags: 
      - Users
      description: Log into server.
      requestBody:
        description: Data the user provides for authentication.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/logInUser
      responses:
        '200':
          description: User logged in successfully.
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/loggedInUser
        '401':
          description: Incorrect login credentials
        '404':
          description: Not Found
  /users/register:
    summary: Registration of a user account

    post:
      tags: 
      - Users
      description: Create a new user.
      requestBody:
        description: Data that makes up a user.
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json#/definitions/createUser
      responses:
        '201':
          description: Created user
          content:
            application/json:
              schema:
                $ref: ./user.json#/definitions/createdUser
...